Add avro serialization

+++ blockchainproducer/src/main/java/com/zenika/kafkaeth/domain/TransactionAdapter.java
@@ -8,16 +8,16 @@ import java.math.BigDecimal;
 public class TransactionAdapter {
     public static com.zenika.kafkaeth.domain.Transaction toTransaction(Transaction tx) {
         return new com.zenika.kafkaeth.domain.Transaction(
-                tx.getNonce(),
+                tx.getNonce().longValue(),
                 tx.getHash(),
                 tx.getBlockHash(),
-                tx.getBlockNumber(),
-                tx.getTransactionIndex(),
+                tx.getBlockNumber().longValue(),
+                tx.getTransactionIndex().longValue(),
                 tx.getFrom(),
                 tx.getTo(),
-                Convert.fromWei(new BigDecimal(tx.getValue()), Convert.Unit.ETHER),
-                tx.getGasPrice(),
-                tx.getGas()
+                Convert.fromWei(new BigDecimal(tx.getValue()), Convert.Unit.ETHER).doubleValue(),
+                tx.getGasPrice().longValue(),
+                tx.getGas().longValue()
         );
     }
 }
